requirements.txt
anthropic>=0.40.0
openai>=1.50.0
streamlit>=1.36.0
python-dotenv>=1.0.0
